[ACM 10232, 10274] Keep managedclusteraddon for local-cluster  (#1370)
* keep managedclusteraddon for local-cluster


* comment


* comment


* refactor


---------

+++ operators/multiclusterobservability/controllers/placementrule/placementrule_controller.go
@@ -196,10 +196,6 @@ func (r *PlacementRuleReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 			log.Error(err, "Failed to delete observabilityaddon")
 			return ctrl.Result{}, err
 		}
-		err = deleteManagedClusterRes(r.Client, localClusterName)
-		if err != nil {
-			return ctrl.Result{}, err
-		}
 	}
 	if operatorconfig.IsMCOTerminating {
 		delete(managedClusterList, "local-cluster")
@@ -270,7 +266,7 @@ func (r *PlacementRuleReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 	// but the managedclusteraddon for observability will not deleted by the cluster manager, so check against the
 	// managedclusteraddon list to remove the managedcluster resources after the managedcluster is detached.
 	for _, mcaddon := range managedclusteraddonList.Items {
-		if !slices.Contains(latestClusters, mcaddon.Namespace) {
+		if !slices.Contains(latestClusters, mcaddon.Namespace) && mcaddon.Namespace != config.GetDefaultNamespace() {
 			reqLogger.Info("To delete managedcluster resources", "namespace", mcaddon.Namespace)
 			err = deleteManagedClusterRes(r.Client, mcaddon.Namespace)
 			if err != nil {
@@ -298,6 +294,12 @@ func (r *PlacementRuleReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 	}
 
 	if deleteAll {
+		// delete managedclusteraddon for local-cluster
+		err = deleteManagedClusterRes(r.Client, localClusterName)
+		if err != nil {
+			return ctrl.Result{}, err
+		}
+
 		opts.Namespace = ""
 		err = r.Client.List(context.TODO(), workList, opts)
 		if err != nil {
